Introduces multiple DownsamplingIteartors that implement different strategies for downsampling a stream of SAMRecords, including one that works in constant memory while correctly handling secondary and supplemental records, one that attempts to minimize the difference between the actual emitted proportion and the requested proportion, and one that is a hybrid of the two - trying for lower but not fixed memory usage while increasing accuracy.
Included is a Strategy enum to describe the strategies used, and a factory for creating downsampling iterators.

Lastly some tests are included.  The attempt to test the approximate accuracy of the downsamplers, but they are a compromise between runtime and accuracy (accuracy is increased by using more reads, but so is test runtime).